requirements.txt
streamlit
chromadb
sentence-transformers
PyMuPDF
spacy
torch
huggingface-hub
groq
python-dotenv
langsmith
